renamed waiver types to admin types
