+++ frontend/src/components/pages/GlobalForms/Footer.tsx
@@ -6,7 +6,7 @@ import {
   useDisclosure,
 } from "@chakra-ui/react";
 import React from "react";
-import { WaiverClause } from "../../../types/WaiverTypes";
+import { WaiverClause } from "../../../types/AdminTypes";
 import WaiverAddSectionModal from "./WaiverAddSectionModal";
 
 type FooterProps = {
